SOC-7484 misc: bcm-vk: use proper return type of long for ioctl handler
Change return value of ioctl from int to long so it is the proper
return type.

Reviewed-on: http://gerrit-ccxsw.broadcom.net/66096
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>

+++ drivers/misc/bcm-vk/bcm_vk.c
@@ -33,7 +33,7 @@ static irqreturn_t bcm_vk_irqhandler(int irq, void *dev_id)
 
 static long bcm_vk_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
 {
-	int ret = -EINVAL;
+	long ret = -EINVAL;
 	return ret;
 }
